CR-1071 refactor based on Phil's comments
