--- src/test/resources/pgp/key1.pub
@@ -1,30 +1,0 @@
------BEGIN PGP PUBLIC KEY BLOCK-----
-
-mQENBF9aT7UBCADadow8ZQPKs0ewpXAOJgP4W1S2vz1Qazwywf9n8yJmOe50XUMj
-5PNLYC587NE/rzXSecL/49yetNT1C4QJbAOhUd3lBwslA6BJHX+fEvJyG82L4pmn
-XPKK0Jtu9fqMlEyHPy/Y539QK0sHwLCPOMRcsYjPDNVB7VJIppjqmuwUCnFBKP75
-6Na5u5SMRsYYdpkLPr7wdMB57/WtTJWls57s/undZw2+CxjBXWIfZ318P6zweZXt
-dIaqkWV71kk1DkobOLwBBxqkj+cH+D+pJOHr5wIQvsbfl14KF2ZHb/ibU2z3egYW
-ezdUYTMnwAfJLh33505ikDpiiNFfvuz0c6FjABEBAAG0HlJvYiBDYXRsaW5nIDxn
-cGdAcnBjLXNvZnQuY29tPokBVAQTAQgAPhYhBFcS5UyvVVrO7MPSaN7ePZdLjU2h
-BQJfWk+1AhsDBQkDwmcABQsJCAcCBhUKCQgLAgQWAgMBAh4BAheAAAoJEN7ePZdL
-jU2h/Y8IAK2GR+WFASoISR/0cr11niPTa8dw3IAGFHj731OoUqactlkL0t1fSQ0k
-oCcLFr76U5K0dxtdLGvDc+hGdUqNONcBYDxLLGDZDpdkigblO7ppYM0GpyMQxdcX
-FIrBICSzoulW9fAMQPAxaboUGbSQSGfUFxEaXf0npKffIxN6Rz7+IA2d+bsPZOSq
-sSaVjd1C3agx+N0uEAgv9OU9mlGoVa7mMvzeprltqRQHl6DAGMpDBwpbVeYrsw7g
-QOwWXs6lAnuw9WnuZiyIFAZrgZtVEM8c0KLsSMToQMkhst0L3Hfp7i0vjkdY4awr
-GhC87S4HadDwS/kRmGUBuLM6MyeJnDm5AQ0EX1pPtQEIAMLAfzwb4KB3k0NgKDOB
-RG06J4yG9WI7DJch6m55oXTn6TOStgl1jdbn5SdevhJo2A2iQjiYCHSPOZV93kv7
-RPEvrHjupx56tIM0FxL06Gm7QDHKKL/+AQm1t4+vGDeRC7VKIxkfUE7EZCkqoYQk
-0g44KoZnnLyr0TZLfYuE4oKAaQRnoQ3nqFgWrqGNJyZvvV0hGXtBYzjkPRHHCBZ2
-tjy8YIZUwOBMHjgVAwmf950ahoN9EdUmRPZwSqs1NLVAYKDu61Gbxur4pkN1qCEz
-PBX1UvTNkw6Jz/vImBf39S/4RwtBs82q06LrhPy95hTj3lGZNjZ8wzYRSHOZLhPT
-+ukAEQEAAYkBPAQYAQgAJhYhBFcS5UyvVVrO7MPSaN7ePZdLjU2hBQJfWk+1AhsM
-BQkDwmcAAAoJEN7ePZdLjU2h57EIALtvASgMV+vNmrFxpiGDfYrgZ7r9jnXrVLiv
-M0/IPMPbvvfpAUkvuFNcJzHVDBaPtbZfOAWAqUnLps3sZQi9i19MYcWezEKnCeyD
-Tpkqy7fx5z412mYgR+MmhvL7nORqHvZZTSziajCQ3a0wWG2dQuj3ETOMHh2hLvmx
-YfuFsYsscePqsE5IABenHhMFH9FcioDnLEC60feq2tZX/novuPrBF6NCfO5SqvcT
-j/lVtejSwTIku1ZWkZ1zuZaluKJXy7oHt//SLj2o6awtYMSbeGQoblTClnqP/urK
-QdQK/+A1rUqZiPdE39usftPUkPd/JZ1RhxHN83gqCeByEt+ySVY=
-=yGOR
------END PGP PUBLIC KEY BLOCK-----
