--- src/test/resources/pgp/key2.pub
@@ -1,30 +1,0 @@
------BEGIN PGP PUBLIC KEY BLOCK-----
-
-mQENBF9elUIBCADqsWADIAyavsPLWIfQ24+uMhz3ziIhwbBKQKwbS7scWbO5sYUz
-0SFPnXINEYlJ4HHOzbRSKplVtxeLMb/ieQCp7yJrhLmKD8NteVqWASPa1UQDqefH
-7ngbc8aJ+tPdrc4qZKWfrLIcvksIIs4kvxwLymRQUbYNvIXNba1GDdmDQ9Jj9sN0
-9aoycTvlR2pCokyIV9z3kRq4o8jT2/JMgNmpPSfSG4FL+bbW7rF9r30nQyYMFXBj
-waemhzqp4bBtRwPFvAkEZmuEBCjneEJy6RYUepnztsj1QONku8knd6/Zk/+yXNbe
-bTljiIOuv70iy6BwR6ScMh4eh7/DbotF3QOBABEBAAG0JUZyZWQgQmxvZ2dzIDxm
-cmVkYmxvZ2dzQHJwYy1zb2Z0LmNvbT6JAVQEEwEIAD4WIQTZ6qhG/MSXWignbQFb
-cpaazuCHAQUCX16VQgIbAwUJA8JnAAULCQgHAgYVCgkICwIEFgIDAQIeAQIXgAAK
-CRBbcpaazuCHAfB7CAC4HcBDCtYGU/mZHRRmJJp/ogZU8ODqcgR787VB3UVRWpfx
-AE1ESZzt7uOFI1jAQ1ryLXFC3hVcyTyOpzc4j95+vgsUF3qdxKDLZY67/oZuH0VH
-goGN2+xMZBx+a3aC7pi3m2kpJXVBSpjbheOGGHDTmAVvW2oP/0gjRoRt4lm/U3V6
-Do0iWtay2DQb7Vd0ztC/lUpXtpwx3P0IfrXgHkg4lqam/Qpy4XkKyUdz1Go2++Ke
-7NWUH84Napa9mgdT7jU9UMDd813DVwLj63Fs1cHpr/b2vjqIg6nNF4YQfhMgVYNY
-t7d/441MG0h5zg0hC5kTHuByYTHkNdR71afZk+w8uQENBF9elUIBCADHaWh/J639
-edv1nkPJHj9DS57XVwgb+PE3ObzyWYXjjSDWjcaUNluNPbGN3QqGM6vj6fZSmja6
-zB8iv6TMhyBGFhDOOApF6uJX0IV5z7FKjoEDBpRZuhD7gLmzZtSbbT5sLk4C7Nzr
-B3YU3mlnHuyDwD0vT6f75A0y7YCaF3U3rNEeakvvkw5wIODBr4mMlb10wN7lIxaM
-k9dCzBbjUyggaWXEH/BvljS5gaAfSHA1lE13471qczcYnhFAZmnHmU+OdoYZN39e
-TxN+3JIt92jVqdTQqRHZrtwH1ai2fBpKtiu8ydjccxHY4YKgGeA2fttTr04cF7Hc
-B3LnKHigVo97ABEBAAGJATwEGAEIACYWIQTZ6qhG/MSXWignbQFbcpaazuCHAQUC
-X16VQgIbDAUJA8JnAAAKCRBbcpaazuCHAeMvB/9FBXx5h0hdl4MS+JUX+Wl4LnxK
-/WP+tVsoeiXfsi3zd/kCpZkMj5txUtE7KY38308cHgrajJeWbwRHLIBarI01BBg3
-pQC1T/U8zL+qVSbMCyfB+JE+CKd16K168zW5P+nd4fNe2m9SI9BInmzvCtNx/4NO
-ewY8wQhw3r2kv5lmruOWNwyAg43d0NXp6d16rw0GGDlXZvlcwnJJLlbyntATiNWg
-y/gUQYWmXl0EZnTpUw6pazOx52QvNuci9O4ovMXVxQXJUGa5vurergQh3dBS2XI+
-Jz9M3FARZQnmCJRzf+RQY/xMv2iVBY1DILbC2kQlcKUrpKKLjRdv4FmGu8sc
-=H6nv
------END PGP PUBLIC KEY BLOCK-----
